abstract existing metrics collection
